+++ Tools/ScriptX/ScriptX/test/src/EngineTest.cc
@@ -70,7 +70,7 @@ TEST_F(EngineTest, SmartPointer) {
   uniquePtr.release();
   uniquePtr1.release();
   engine = nullptr;
-  // destory engine
+  // destroy engine
   sharedPtr.reset();
 }
 
@@ -131,6 +131,30 @@ TEST_F(EngineTest, LuaBuiltIns) {
   }
 }
 
+#endif
+
+#ifdef SCRIPTX_LANG_JAVASCRIPT
+TEST_F(EngineTest, JsPromiseTest) {
+  EngineScope scope(engine);
+
+  int value = 0;
+  auto setValue = Function::newFunction([&value](int val) { value = val; });
+  engine->set("setValue", setValue);
+  engine->eval(
+      u8R"(
+        const promise = new Promise((resolve, reject) => {
+            resolve(1);
+        });
+
+        promise.then(num => {
+            setValue(num);
+        });
+    )");
+
+  engine->messageQueue()->shutdown(true);
+  engine->messageQueue()->loopQueue(utils::MessageQueue::LoopType::kLoopAndWait);
+  EXPECT_EQ(value, 1);
+}
 #endif
 
 }  // namespace script::test